backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
google-generativeai==0.8.0
edge-tts==6.1.12
pydub==0.25.1
python-dotenv==1.0.1
python-multipart==0.0.9
feedgen==1.0.0
